+++ tests/unit/test_options.py
@@ -183,12 +183,13 @@ class OptionsSuite(unittest.TestCase):
             sense=Sense.MIN,
             objective=Sum((i, j), c[i, j] * x[i, j]),
         )
-        # logoption = 2
-        transport.solve(create_log_file=True)
 
-        # logoption = 4
-        with self.assertRaises(NotImplementedError):
-            transport.solve(output=sys.stdout, create_log_file=True)
+        transport.solve()  # logoption=0
+        transport.solve(create_log_file=True)  # logoption = 2
+        transport.solve(output=sys.stdout)  # logoption = 3
+        transport.solve(
+            output=sys.stdout, create_log_file=True
+        )  # logoption = 4
 
         # test logfile
         logfile_name = os.path.join(os.getcwd(), "tmp", "log.txt")